Move all scope validations out of Scope and XmlScope into vaidator
