+++ src/XmlScope.ts
@@ -1,11 +1,6 @@
 import { Scope } from './Scope';
-import { DiagnosticMessages } from './DiagnosticMessages';
 import type { XmlFile } from './files/XmlFile';
-import type { FileReference } from './interfaces';
-import { DiagnosticOrigin } from './interfaces';
 import type { Program } from './Program';
-import type { SGElement } from './parser/SGTypes';
-import { SGFieldTypes } from './parser/SGTypes';
 import util from './util';
 import { SymbolTypeFlag } from './SymbolTypeFlag';
 import type { BscFile } from './files/BscFile';
@@ -79,109 +74,6 @@ export class XmlScope extends Scope {
             }
         }
         return result;
-    }
-
-    protected _validate() {
-        //validate brs files
-        super._validate();
-
-        //detect when the child imports a script that its ancestor also imports
-        this.diagnosticDetectDuplicateAncestorScriptImports();
-
-        //validate component interface
-        this.diagnosticValidateInterface();
-    }
-
-    private diagnosticValidateInterface() {
-        if (!this.xmlFile.parser.ast?.componentElement?.interfaceElement) {
-            return;
-        }
-        const iface = this.xmlFile.parser.ast.componentElement.interfaceElement;
-        const callableContainerMap = this.getCallableContainerMap();
-        //validate functions
-        for (const func of iface.functions) {
-            const name = func.name;
-            if (!name) {
-                this.diagnosticMissingAttribute(func, 'name');
-            } else if (!callableContainerMap.has(name.toLowerCase())) {
-                this.diagnostics.push({
-                    ...DiagnosticMessages.xmlFunctionNotFound(name),
-                    range: func.getAttribute('name')?.tokens.value.range,
-                    file: this.xmlFile,
-                    origin: DiagnosticOrigin.Scope
-                });
-            }
-        }
-        //validate fields
-        for (const field of iface.fields) {
-            const { id, type, onChange } = field;
-            if (!id) {
-                this.diagnosticMissingAttribute(field, 'id');
-            }
-            if (!type) {
-                if (!field.alias) {
-                    this.diagnosticMissingAttribute(field, 'type');
-                }
-            } else if (!SGFieldTypes.includes(type.toLowerCase())) {
-                this.diagnostics.push({
-                    ...DiagnosticMessages.xmlInvalidFieldType(type),
-                    range: field.getAttribute('type')?.tokens.value.range,
-                    file: this.xmlFile,
-                    origin: DiagnosticOrigin.Scope
-                });
-            }
-            if (onChange) {
-                if (!callableContainerMap.has(onChange.toLowerCase())) {
-                    this.diagnostics.push({
-                        ...DiagnosticMessages.xmlFunctionNotFound(onChange),
-                        range: field.getAttribute('onchange')?.tokens.value.range,
-                        file: this.xmlFile,
-                        origin: DiagnosticOrigin.Scope
-                    });
-                }
-            }
-        }
-    }
-
-    private diagnosticMissingAttribute(tag: SGElement, name: string) {
-        this.diagnostics.push({
-            ...DiagnosticMessages.xmlTagMissingAttribute(tag.tokens.startTagName.text, name),
-            range: tag.tokens.startTagName.range,
-            file: this.xmlFile,
-            origin: DiagnosticOrigin.Scope
-        });
-    }
-
-    /**
-     * Detect when a child has imported a script that an ancestor also imported
-     */
-    private diagnosticDetectDuplicateAncestorScriptImports() {
-        if (this.xmlFile.parentComponent) {
-            //build a lookup of pkg paths -> FileReference so we can more easily look up collisions
-            let parentScriptImports = this.xmlFile.getAncestorScriptTagImports();
-            let lookup = {} as Record<string, FileReference>;
-            for (let parentScriptImport of parentScriptImports) {
-                //keep the first occurance of a pkgPath. Parent imports are first in the array
-                if (!lookup[parentScriptImport.destPath]) {
-                    lookup[parentScriptImport.destPath] = parentScriptImport;
-                }
-            }
-
-            //add warning for every script tag that this file shares with an ancestor
-            for (let scriptImport of this.xmlFile.scriptTagImports) {
-                let ancestorScriptImport = lookup[scriptImport.destPath];
-                if (ancestorScriptImport) {
-                    let ancestorComponent = ancestorScriptImport.sourceFile as XmlFile;
-                    let ancestorComponentName = ancestorComponent.componentName?.text ?? ancestorComponent.destPath;
-                    this.diagnostics.push({
-                        file: this.xmlFile,
-                        range: scriptImport.filePathRange,
-                        ...DiagnosticMessages.unnecessaryScriptImportInChildFromParent(ancestorComponentName),
-                        origin: DiagnosticOrigin.Scope
-                    });
-                }
-            }
-        }
     }
 
     public getAllFiles() {
